[fix] card_207 顔認識ぼかし機能
* extensionをfileprivateに変更
